Abstract stat box into its own component
