+++ src/components/ProfileOverview/ProfileOverview.js
@@ -1,16 +1,8 @@
 import React from 'react';
 import * as S from './ProfileOverview.styles';
 import Type from '../common/Type';
-import StatBar from '../StatBar';
-
-const statKeyMap = {
-  hp: 'hp',
-  attack: 'atk',
-  defense: 'def',
-  'special-attack': 'sp. atk',
-  'special-defense': 'sp. def',
-  speed: 'spd',
-};
+import SummaryBox from '../SummaryBox';
+import StatsBox from '../StatsBox';
 
 export default function ProfileOverview({
   id = 0,
@@ -36,22 +28,13 @@ export default function ProfileOverview({
         <S.Image src={artworkUrl}/>
       </S.Header>
       <S.ContentGrid>
-        <S.StatsBox>
-          {stats.map((stat) => (
-            <S.StatRow key={stat.key}>
-              <S.StatKey>
-                {statKeyMap[stat.key]}
-              </S.StatKey>
-              <StatBar stat={stat.value}/>
-              <S.StatValue>
-                {stat.value}
-              </S.StatValue>
-            </S.StatRow>
-          ))}
-        </S.StatsBox>
-        <S.SummaryBox>
-          summary
-        </S.SummaryBox>
+        <StatsBox stats={stats}/>
+        <SummaryBox
+          abilities={abilities}
+          height={height}
+          weight={weight}
+          baseExperience={baseExperience}
+        />
         <S.SummaryBox>
           <S.Subtitle>Abilities</S.Subtitle>
           <S.Abilities>
