+++ src/components/StatsBox/index.js
@@ -1,0 +1,3 @@
+import StatsBox from './StatsBox';
+
+export default StatsBox;
