+++ src/components/SummaryBox/SummaryBox.styles.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { Styles } from '../../themes';
 
-export const Wrapper = styled(Box)`
+export const Wrapper = styled(Styles.Box)`
 
 `;
